upgraded KCL
