+++ src/main/resources/reference.conf
@@ -652,6 +652,70 @@ kinesis {
          # Default: no timeout
          #timeoutInSeconds =
 
+         # The amount of milliseconds to wait before graceful shutdown forcefully terminates the process.
+         # Default: 5000
+         #shutdownGraceMillis = 5000
+
+         # Internally support timeouts and retries for GetRecords calls.
+         #
+         # If retryGetRecordsInSeconds is set
+         # And maxGetRecordsThreadPool is set
+         # Then getRecords will asynchronously retry internally using a CompletionService of
+         # max size "maxGetRecordsThreadPool" with "retryGetRecordsInSeconds" between each retry.
+         #
+         # NOTE: this enables the AsynchronousGetRecordsRetrievalStrategy for getRecords
+         #
+         # Time in seconds to wait before the worker retries to get a record
+         # Default: Optional value, default not set
+         #retryGetRecordsInSeconds = 1
+         #
+         # max number of threads in the getRecords thread pool (per shard)
+         # Default: Optional value, default not set
+         #maxGetRecordsThreadPool = 2
+
+
+         #######################
+         # Pre-fetching config #
+         #######################
+         # Pre-fetching will retrieve and queue additional records from Kinesis while the
+         # application is processing existing records.
+         # Pre-fetching can be enabled by setting dataFetchingStrategy to PREFETCH_CACHED. Once
+         # enabled an additional fetching thread will be started to retrieve records from Kinesis.
+         # Retrieved records will be held in a queue until the application is ready to process them.
+
+         # Which data fetching strategy to use (DEFAULT, PREFETCH_CACHED)
+         # Default: DEFAULT
+         dataFetchingStrategy = PREFETCH_CACHED
+
+         #
+         # Pre-fetching supports the following configuration values:
+         #
+
+         # The maximum number of process records input that can be queued
+         # Default: 3
+         #maxPendingProcessRecordsInput = 3
+
+         # The maximum number of bytes that can be queued
+         # Default 8388608 (8 * 1024 * 1024 / 8Mb)
+         #maxCacheByteSize = 8388608
+
+         # The maximum number of records that can be queued
+         # Default: 30000
+         #maxRecordsCount = 30000
+
+         # The amount of time to wait between calls to Kinesis
+         # Default: 1500
+         #idleMillisBetweenCalls = 1500
+
+         ##############################
+         # End of Pre-fetching config #
+         ##############################
+
+
+         # Milliseconds after which the logger will log a warning message for the long running task
+         # Default: not set
+         #logWarningForTaskAfterMillis = 100
+
       }
    }
 }
